op-devstack/sysgo: Remove interopgen implementation.
It was unused and created confusion about how deployments were handled.
Move types used by deployer to more appropriate files.
